fix(vscode): test the tool cards through the panel harness
The previous commit moved the tool-card helpers into a second browser
script, media/tool-card-utils.js, because chat-panel.js had no test
harness. Rebasing onto main invalidated that: #867 added
source/vscode/chat-panel-thoughts.spec.ts, which boots chat-panel.js in
a VM against a stub DOM. It loads only chat-panel.js, so the new
top-level `globalThis.NanocoderToolCardUtils` destructure threw and took
all 13 of its tests down with it.

Fold the helpers back into chat-panel.js and drop the extra script, its
tag, and the provider wiring - the reason for splitting them out no
longer holds, and a load-order dependency between two media scripts is
worth avoiding on its own.

Test them through that harness instead, which reaches what the extracted
module never could: the card is now driven by real ACP updates and the
rendered DOM is asserted on, covering the queued label and the gated
"Open Diff" affordance end to end.

To share the harness it moves to source/vscode/chat-panel-harness.ts,
unchanged apart from three gaps that the tool-card tests exposed:

- querySelector/querySelectorAll returned null and []. The panel reaches
  into its own cards with them, so those branches were never exercised.
  They now resolve single-class and tag selectors.
- className and classList were independent stores, so setting one and
  reading the other silently disagreed - assertions either way passed
  vacuously. They now share one set, as they do in a DOM.
- postMessage was a no-op, so nothing the panel sent to the extension
  host could be asserted on. It is recorded and exposed as `sent`.
