+++ plugins/vscode/media/chat-panel.html
@@ -96,7 +96,6 @@
 			</div>
 		</div>
 	<script nonce="{{nonce}}" src="{{markedUri}}"></script>
-	<script nonce="{{nonce}}" src="{{toolCardUtilsUri}}"></script>
 	<script nonce="{{nonce}}" src="{{scriptUri}}"></script>
 </body>
 </html>
