+++ source/vscode/chat-panel-thoughts.spec.ts
@@ -1,256 +1,5 @@
-import {readFileSync} from 'node:fs';
-import {fileURLToPath} from 'node:url';
-import {createContext, runInContext} from 'node:vm';
 import test from 'ava';
-
-const PANEL_SOURCE = readFileSync(
-	fileURLToPath(
-		new URL('../../plugins/vscode/media/chat-panel.js', import.meta.url),
-	),
-	'utf8',
-);
-
-const SHELL_IDS = [
-	'add-image-btn',
-	'attach-btn',
-	'chat-input',
-	'chat-view',
-	'close-modal-btn',
-	'composer-box',
-	'context-chips',
-	'history-list',
-	'history-view',
-	'icon-send',
-	'icon-stop',
-	'image-modal',
-	'image-preview-container',
-	'image-upload',
-	'messages-container',
-	'modal-image',
-	'mode-dropdown',
-	'mode-trigger',
-	'mode-trigger-label',
-	'model-dropdown',
-	'model-trigger',
-	'model-trigger-label',
-	'provider-dropdown',
-	'provider-trigger',
-	'provider-trigger-label',
-	'send-stop-btn',
-];
-
-type StubElement = any;
-
-function createElement(tagName: string): StubElement {
-	const classes = new Set<string>();
-	const attributes = new Map<string, string>();
-	let html = '';
-	let text = '';
-
-	const element: StubElement = {
-		tagName,
-		id: '',
-		className: '',
-		title: '',
-		value: '',
-		disabled: false,
-		files: null,
-		style: {},
-		dataset: {},
-		children: [] as StubElement[],
-		parentElement: null as StubElement | null,
-		classList: {
-			add: (...names: string[]) => names.forEach(name => classes.add(name)),
-			remove: (...names: string[]) =>
-				names.forEach(name => classes.delete(name)),
-			contains: (name: string) => classes.has(name),
-			toggle: (name: string, force?: boolean) => {
-				const on = force === undefined ? !classes.has(name) : force;
-				if (on) classes.add(name);
-				else classes.delete(name);
-			},
-		},
-		appendChild(child: StubElement) {
-			child.parentElement = element;
-			element.children.push(child);
-			return child;
-		},
-		removeChild(child: StubElement) {
-			element.children = element.children.filter(
-				(candidate: StubElement) => candidate !== child,
-			);
-			child.parentElement = null;
-			return child;
-		},
-		remove() {
-			element.parentElement?.removeChild(element);
-		},
-		querySelector: () => null,
-		querySelectorAll: () => [],
-		closest: () => null,
-		setAttribute: (name: string, value: string) => attributes.set(name, value),
-		getAttribute: (name: string) => attributes.get(name) ?? null,
-		addEventListener: () => {},
-		removeEventListener: () => {},
-		focus: () => {},
-		click: () => {},
-		scrollTop: 0,
-		scrollHeight: 0,
-	};
-
-	Object.defineProperty(element, 'innerHTML', {
-		get: () => html,
-		set: (value: string) => {
-			html = String(value);
-			element.children = [];
-		},
-	});
-	Object.defineProperty(element, 'textContent', {
-		get: () => text,
-		set: (value: string) => {
-			text = String(value);
-		},
-	});
-
-	return element;
-}
-
-function findById(root: StubElement, id: string): StubElement | null {
-	for (const child of root.children) {
-		if (child.id === id) return child;
-		const found = findById(child, id);
-		if (found) return found;
-	}
-	return null;
-}
-
-function createPanel(options: {marked?: boolean} = {}) {
-	const clock = {now: 1_700_000_000_000};
-
-	class FakeDate extends Date {
-		constructor(...args: any[]) {
-			super(...((args.length ? args : [clock.now]) as []));
-		}
-		static now() {
-			return clock.now;
-		}
-	}
-
-	let nextTimerId = 1;
-	const timers = new Map<number, {fn: () => void; repeat: boolean}>();
-	const schedule = (fn: () => void, repeat: boolean) => {
-		const id = nextTimerId++;
-		timers.set(id, {fn, repeat});
-		return id;
-	};
-	const cancel = (id: number) => {
-		timers.delete(id);
-	};
-
-	const root = createElement('html');
-	const body = createElement('body');
-	root.appendChild(body);
-	for (const id of SHELL_IDS) {
-		const element = createElement('div');
-		element.id = id;
-		body.appendChild(element);
-	}
-
-	const messageListeners: ((event: {data: unknown}) => void)[] = [];
-	const sandbox: Record<string, unknown> = {
-		document: {
-			body,
-			createElement,
-			createElementNS: (_namespace: string, tagName: string) =>
-				createElement(tagName),
-			getElementById: (id: string) => findById(root, id),
-			querySelector: () => null,
-			querySelectorAll: () => [],
-			addEventListener: () => {},
-		},
-		window: {
-			addEventListener: (
-				type: string,
-				fn: (event: {data: unknown}) => void,
-			) => {
-				if (type === 'message') messageListeners.push(fn);
-			},
-		},
-		navigator: {userAgent: '', clipboard: {writeText: async () => {}}},
-		acquireVsCodeApi: () => ({
-			postMessage: () => {},
-			getState: () => undefined,
-			setState: () => {},
-		}),
-		setTimeout: (fn: () => void) => schedule(fn, false),
-		setInterval: (fn: () => void) => schedule(fn, true),
-		clearTimeout: cancel,
-		clearInterval: cancel,
-		Date: FakeDate,
-		console,
-	};
-	if (options.marked) {
-		sandbox.marked = {parse: (value: string) => `<md>${value}</md>`};
-	}
-
-	createContext(sandbox);
-	runInContext(PANEL_SOURCE, sandbox);
-
-	const container = findById(root, 'messages-container') as StubElement;
-
-	return {
-		container,
-		post(message: unknown) {
-			for (const listener of messageListeners) listener({data: message});
-		},
-		update(update: Record<string, unknown>) {
-			this.post({type: 'acpUpdate', update});
-		},
-		thought(value: string) {
-			this.update({
-				sessionUpdate: 'agent_thought_chunk',
-				content: {type: 'text', text: value},
-			});
-		},
-		text(value: string) {
-			this.update({
-				sessionUpdate: 'agent_message_chunk',
-				content: {type: 'text', text: value},
-			});
-		},
-		tool(toolCallId: string) {
-			this.update({
-				sessionUpdate: 'tool_call',
-				toolCallId,
-				title: 'read_file',
-				status: 'pending',
-			});
-		},
-		finish() {
-			this.update({sessionUpdate: 'prompt_response'});
-		},
-		userMessage(value: string) {
-			this.update({
-				sessionUpdate: 'user_message_chunk',
-				content: {type: 'text', text: value},
-			});
-		},
-		advance(ms: number) {
-			clock.now += ms;
-		},
-		runTimers() {
-			for (const [id, timer] of [...timers]) {
-				if (!timer.repeat) timers.delete(id);
-				timer.fn();
-			}
-		},
-		boxes(): StubElement[] {
-			return container.children.filter((child: StubElement) =>
-				child.className.includes('thought-aggregator'),
-			);
-		},
-	};
-}
+import {createPanel, type StubElement} from '@/vscode/chat-panel-harness';
 
 const titleOf = (box: StubElement) => box.children[0].children[0].textContent;
 const bodyOf = (box: StubElement) => box.children[1].textContent;
